rename package from models to graphs
